django_2/productUser/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers 
pika==1.1.0 
psycopg2-binary
celery
django-crontab==0.7.1

